+++ .gitignore
@@ -35,4 +35,4 @@ build/
 ### Mac OS ###
 .DS_Store
 
-/config.properties
+/src/main/resources/config.properties
